refactor: mirror src/extensions/ structure to user-side namespace
Move src/ui/registry.ts and src/extensions/loader.ts into
src/extensions/{pi,ui}/ so source layout matches what user sees at
~/.isotopes/extensions/{pi,ui,channels}/. channels/ gets a placeholder
README documenting the reserved slot.

Plus: fix stale runtime.ts:67 comment that still pointed at the old
~/.isotopes/extensions/ path; update CLAUDE.md to consolidate the two
prior bullets (extensions/ + ui/) into one extensions/ description.
