+++ src/app.ts
@@ -16,8 +16,8 @@ import { HeartbeatManager } from "./automation/heartbeat.js";
 import { getIsotopesHome } from "./paths.js";
 import { AgentRuntime } from "./agent/runtime.js";
 import { runAgent } from "./agent/runtime-adapter.js";
-import { discoverExtensionPaths } from "./extensions/loader.js";
-import { discoverUIEntries } from "./ui/registry.js";
+import { discoverExtensionPaths } from "./extensions/pi/loader.js";
+import { discoverUIEntries } from "./extensions/ui/registry.js";
 import { createDiscordTransport } from "./legacy/discord/index.js";
 import type { Transport } from "./legacy/gateway/types.js";
