+++ src/extensions/channels/README.md
@@ -1,0 +1,5 @@
+# channels/
+
+Reserved slot for future channel adapter discovery (Discord, Feishu, etc.).
+
+Mirrors the user-side path `~/.isotopes/extensions/channels/`. Empty until a channel adapter actually needs runtime discovery — current channels are hard-wired in `app.ts` and `legacy/discord/`.
