+++ src/extensions/ui/registry.ts
@@ -1,7 +1,7 @@
 import fs from "node:fs";
 import path from "node:path";
-import { getIsotopesHome } from "../paths.js";
-import { createLogger } from "../logging/logger.js";
+import { getIsotopesHome } from "../../paths.js";
+import { createLogger } from "../../logging/logger.js";
 
 const log = createLogger("ui");
